requirements.txt
pandas==2.3.3
scikit-learn==1.4.2
mlflow==3.6.0
PyYAML==6.0.3
pytest==9.0.1
requests==2.32.3
numpy>=1.26.4,<2.2
joblib==1.4.2